Crate honeycomb

honeycomb

Honeycomb aims to provide a safe, efficient and scalable implementation of combinatorial maps for meshing applications. More specifically, the goal is to converge towards a (or multiple) structure(s) adapted to algorithms exploiting GPU and many-core architectures.

Structure

This crate acts as the user-facing API, re-exporting components and items implemented in the following sub-crates:

honeycomb-core – core structures implementations

honeycomb-kernels – algorithm implementations

honeycomb-render – visual debugging tool

Features

Two features can be enabled to control which implementations are exposed:

kernels – content from the honeycomb-kernels crate

render – content from the honeycomb-render crate

Note that:

the kernels feature is enabled by default since it does not require any more dependencies than the core crate.

the render feature is disabled by default; enabling it significantly lengthen the dependency tree as well as the compilation time.

Quickstart

For usage examples, refer to examples hosted in the repository; there also are documentation examples for important items:

CMap2

CMapBuilder

grisubal

Re-exports

pub use honeycomb_core as core;

pub use honeycomb_kernels as kernels;

pub use honeycomb_render as render;

Modules

prelude

commonly used items