honeycomb::prelude

Module cell_insertion

Available on crate feature kernels only.

Cell insertion functions

This module contains implementations of cell insertion methods. Due to definition issues, we only implement insertion of N-1-cell into N-cell, for example vertex (0-cell) in edge (1-cell).

Enums

VertexInsertionError

Error-modeling enum for vertex insertion routines.

Functions

insert_vertex_on_edge

Insert a vertex in an edge, cutting it into two segments.

insert_vertices_on_edge

Insert n vertices in an edge, cutting it into n+1 segments.