honeycomb::prelude
Module splits

Cell splitting functions

This module contains implementations of cell splitting methods. We currently define two edge-splitting methods, depending on the number of splits done. Both functions have “no-alloc” variants: these take additional darts as argument in order not to allocate darts during the process.

Enums

SplitEdgeError

Error-modeling enum for edge-splitting routines.

Functions

split_edge

Split an edge into two segments.

split_edge_noalloc

Split an edge into two segments.

splitn_edge

Split an edge into n segments.

splitn_edge_no_alloc

Split an edge into n segments.