honeycomb::prelude
Trait AttributeBind

pub trait AttributeBind:
    Sized
    + Debug
    + Any {
    type StorageType: AttributeStorage<Self>;
    type IdentifierType: From<u32> + ToPrimitive + Clone;

    const BIND_POLICY: OrbitPolicy;
}

Generic attribute trait for support description

This trait can be implemented for a given attribute in order to hint at which components of the map the attribute is bound.

Example

Using the same context as the for the AttributeUpdate example, we can associate temperature to faces if we’re modeling a 2D mesh:

use honeycomb_core::prelude::{AttributeBind, AttributeUpdate, FaceIdentifier, OrbitPolicy};
use honeycomb_core::attributes::AttrSparseVec;

#[derive(Clone, Copy, Debug, PartialEq)]
pub struct Temperature {
    pub val: f32
}

impl AttributeBind for Temperature {
    type StorageType = AttrSparseVec<Self>;
    type IdentifierType = FaceIdentifier;
    const BIND_POLICY: OrbitPolicy = OrbitPolicy::Face;
}

Required Associated Constants

const BIND_POLICY: OrbitPolicy

OrbitPolicy determining the kind of topological entity to which the attribute is associated.

Required Associated Types

type StorageType: AttributeStorage<Self>

Storage type used for the attribute.

type IdentifierType: From<u32> + ToPrimitive + Clone

Identifier type of the entity the attribute is bound to.

Object Safety

This trait is not object safe.