honeycomb::prelude
Trait AttributeUpdate

pub trait AttributeUpdate: Sized {
    // Required methods
    fn merge(attr1: Self, attr2: Self) -> Self;
    fn split(attr: Self) -> (Self, Self);

    // Provided methods
    fn merge_incomplete(attr: Self) -> Self { ... }
    fn merge_from_none() -> Option<Self> { ... }
}

Generic attribute trait for logical behavior description

This trait can be implemented for a given attribute in order to define the behavior to follow when (un)sewing operations result in an update of the attribute.

Example

For an intensive property of a system (e.g. a temperature), an implementation would look like this:

use honeycomb_core::prelude::AttributeUpdate;

#[derive(Clone, Copy, Debug, PartialEq)]
pub struct Temperature {
    pub val: f32
}

impl AttributeUpdate for Temperature {
    fn merge(attr1: Self, attr2: Self) -> Self {
        Temperature { val: (attr1.val + attr2.val) / 2.0 }
    }

    fn split(attr: Self) -> (Self, Self) {
        (attr, attr)
    }

    fn merge_incomplete(attr: Self) -> Self {
        Temperature { val: attr.val / 2.0 }
    }

    fn merge_from_none() -> Option<Self> {
        Some(Temperature { val: 0.0 })
    }
}

let t1 = Temperature { val: 273.0 };
let t2 = Temperature { val: 298.0 };

let t_new = AttributeUpdate::merge(t1, t2); // use AttributeUpdate::_
let t_ref = Temperature { val: 285.5 };

assert_eq!(Temperature::split(t_new), (t_ref, t_ref)); // or Temperature::_

Required Methods

fn merge(attr1: Self, attr2: Self) -> Self

Merging routine, i.e. how to obtain the new attribute value from the two existing ones.

fn split(attr: Self) -> (Self, Self)

Splitting routine, i.e. how to obtain the two attributes from a single one.

Provided Methods

fn merge_incomplete(attr: Self) -> Self

Fallback merging routine, i.e. how to obtain the new attribute value from a single existing value.

The default implementation simply returns the passed value.

fn merge_from_none() -> Option<Self>

Fallback merging routine, i.e. how to obtain the new attribute value from no existing value.

The default implementation return None.

Object Safety

This trait is not object safe.