honeycomb::prelude
Trait CoordsFloat

pub trait CoordsFloat:
    Float
    + Default
    + AddAssign
    + SubAssign
    + MulAssign
    + DivAssign
    + Debug
    + 'static { }

Common trait implemented by types used for coordinate representation.

The static lifetime is a requirements induced by specific implementations that use TypeId; This is used in order to identify types in two contexts:

Interacting with VTK files (io feature),

Coding vertices and generic attributes handling

Object Safety

This trait is not object safe.