honeycomb
Module prelude

commonly used items

This module contains all items commonly used to write a program using combinatorial maps. These items are re-exported from their original crates for ease of use and should cover all basic use cases.

Modules

grisubal

GRISUBAL algorithm description & implementation

splits

Cell splitting functions

triangulation

Polygon triangulation functions

Structs

App

Default render structure.

CMap2

Main map object.

CMapBuilder

Combinatorial map builder structure.

GridDescriptor

Grid description used to generate maps via the map builder.

Orbit2

Generic 2D orbit implementation

Vector2

2D vector representation

Vertex2

2D vertex representation

Enums

BuilderError

Builder-level error enum

CoordsError

Coordinates-level error enum

OrbitPolicy

Orbit search policy enum.

Constants

NULL_DART_ID

Null dart ID value.

NULL_EDGE_ID

Null edge ID value.

NULL_FACE_ID

Null face ID value.

NULL_VERTEX_ID

Null vertex ID value.

NULL_VOLUME_ID

Null volume ID value.

Traits

AttributeBind

Generic attribute trait for support description

AttributeUpdate

Generic attribute trait for logical behavior description

CoordsFloat

Common trait implemented by types used for coordinate representation.

Type Aliases

DartIdType

Dart ID representation type.

EdgeIdType

Edge ID representation type.

FaceIdType

Face ID representation type.

VertexIdType

Vertex ID representation type.

VolumeIdType

Volume ID representation type.