honeycomb

Module prelude

commonly used items

This module contains all items commonly used to write a program using combinatorial maps. These items are re-exported from their original crates for ease of use and should cover all basic use cases.

Modules

cell_insertion

Cell insertion functions

grisubal

grisubal algorithm description & implementation

triangulation

Polygon triangulation functions

Structs

CMap2

2D combinatorial map implementation

CMapBuilder

Combinatorial map builder structure

GridDescriptor

Grid description used by the map builder

Vector2

2D vector structure

Vertex2

2D vertex structure

Enums

BuilderError

Builder-level error enum

CoordsError

Coordinates-level error enum

OrbitPolicy

Orbit search policy enum

Constants

NULL_DART_ID

Null dart ID value

NULL_EDGE_ID

Null edge ID value

NULL_FACE_ID

Null face ID value

NULL_VERTEX_ID

Null vertex ID value

NULL_VOLUME_ID

Null volume ID value

Traits

AttributeBind

Generic attribute trait

AttributeUpdate

Generic attribute trait

CoordsFloat

Generic FP type trait

Functions

render_2d_map

Main rendering function.

Type Aliases

DartIdType

Dart ID representation type

EdgeIdType

Edge ID representation type

FaceIdType

Face ID representation type

VertexIdType

Vertex ID representation type

VolumeIdType

Volume ID representation type