honeycomb::prelude
Trait AttributeUpdate

pub trait AttributeUpdate:
    Sized
    + Send
    + Sync
    + Clone
    + Copy {
    // Required methods
    fn merge(attr1: Self, attr2: Self) -> Self;
    fn split(attr: Self) -> (Self, Self);

    // Provided methods
    fn merge_incomplete(attr: Self) -> Result<Self, CMapError> { ... }
    fn merge_from_none() -> Result<Self, CMapError> { ... }
    fn split_from_none() -> Result<(Self, Self), CMapError> { ... }
}

Generic attribute trait for logical behavior description

This trait can be implemented for a given attribute in order to define the behavior to follow when (un)sewing operations result in an update of the attribute.

Example

For an intensive property of a system (e.g. a temperature), an implementation would look like this:

use honeycomb_core::prelude::{AttributeUpdate, CMapResult};

#[derive(Clone, Copy, Debug, PartialEq)]
pub struct Temperature {
    pub val: f32
}

impl AttributeUpdate for Temperature {
    fn merge(attr1: Self, attr2: Self) -> Self {
        Temperature { val: (attr1.val + attr2.val) / 2.0 }
    }

    fn split(attr: Self) -> (Self, Self) {
        (attr, attr)
    }

    fn merge_incomplete(attr: Self) -> CMapResult<Self> {
        Ok(Temperature { val: attr.val / 2.0 })
    }

    fn merge_from_none() -> CMapResult<Self> {
        Ok(Temperature { val: 0.0 })
    }
}

let t1 = Temperature { val: 273.0 };
let t2 = Temperature { val: 298.0 };

let t_new = AttributeUpdate::merge(t1, t2); // use AttributeUpdate::_
let t_ref = Temperature { val: 285.5 };

assert_eq!(Temperature::split(t_new), (t_ref, t_ref)); // or Temperature::_

Required Methods

fn merge(attr1: Self, attr2: Self) -> Self

Merging routine, i.e. how to obtain the new attribute value from the two existing ones.

fn split(attr: Self) -> (Self, Self)

Splitting routine, i.e. how to obtain the two attributes from a single one.

Provided Methods

fn merge_incomplete(attr: Self) -> Result<Self, CMapError>

Fallback merging routine, i.e. how to obtain the new attribute value from a single existing value.

The returned value directly affects the behavior of UnknownAttributeStorage::merge, UnknownAttributeStorage::try_merge, therefore of sewing methods too.

For example, if this method returns an error for a given attribute, the try_merge method will fail. This allow the user to define some attributes as essential (fail if the merge isn’t done properly from two values) and other as mores flexible (can fallback to a default value).

Errors

The default implementation simply returns the passed value.

fn merge_from_none() -> Result<Self, CMapError>

Fallback merging routine, i.e. how to obtain the new attribute value from no existing value.

The returned value directly affects the behavior of UnknownAttributeStorage::merge, UnknownAttributeStorage::try_merge, therefore of sewing methods too.

For example, if this method returns an error for a given attribute, the try_merge method will fail. This allow the user to define some attributes as essential (fail if the merge isn’t done properly from two values) and others as more flexible (can fallback to a default value).

Errors

The default implementation return Err(CMapError::FailedAttributeMerge).

fn split_from_none() -> Result<(Self, Self), CMapError>

Fallback splitting routine, i.e. how to obtain the new attribute value from no existing value.

The returned value directly affects the behavior of UnknownAttributeStorage::split, UnknownAttributeStorage::try_split, therefore of sewing methods too.

For example, if this method returns an error for a given attribute, the try_split method will fail. This allow the user to define some attributes as essential (fail if the split isn’t done properly from a value) and others as more flexible (can fallback to a default value).

Errors

The default implementation return Err(CMapError::FailedAttributeSplit).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.